Allow override of statement

+++ Dao/Db.php
@@ -381,21 +381,21 @@ class Db implements InterfaceDao {
 		$primaryKey = $this->getPrimaryKey();
 
 		// determine if we insert or update the data
-		if(property_exists($vo,$primaryKey)&&$vo->{$primaryKey}!=''){
+		if(property_exists($vo,$primaryKey)&&$vo->{$primaryKey}!='' && is_null($statement) ){
 			$statement = is_null($statement)?self::STATEMENT_UPDATE:$statement;
 			if(property_exists($vo,'utsUpdate')) $vo->utsUpdate = time();
 
 			$cacheKey = $this->getCache()->getHashKey('DB_GETBYID_'.$this->getVoName(true).'_'.$vo->getId());
 			$this->getCache()->deleteKey($cacheKey);
 		}else{
-			$statement = is_null($statement)?self::STATEMENT_INSERT:$statement;
+			$statement = self::STATEMENT_INSERT;
 			if(property_exists($vo,'utsCreate')) $vo->utsCreate = time();
 		}
 
 		// define each field
 		$set = array();
 		foreach($vo as $property => $value){
-			if($property == $primaryKey) continue;
+			if($property == $primaryKey && $statement === self::STATEMENT_UPDATE ) continue;
 			$set[] = '`'.$property.'` = \''.$this->getConnector()->escape_string($value).'\'';
 		}
